Move getSizeData next to related methods in symbol_size.js
